+++ src/Forms/Components/Concerns/ConfiguresKurdishCalendarPicker.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rawand201\FilamentKurdishCalendar\Forms\Components\Concerns;
+namespace Entensy\FilamentKurdishCalendar\Forms\Components\Concerns;
 
 use Filament\Forms\Components\DateTimePicker;